feat: Moved data models and change imports
