+++ personalization-sdk/src/main/kotlin/com/personalization/OnMessageListener.kt
@@ -1,6 +1,6 @@
 package com.personalization
 
-import com.personalization.features.notification.domain.model.NotificationData
+import com.personalization.sdk.data.models.dto.notification.NotificationData
 
 fun interface OnMessageListener {
     fun onMessage(data: NotificationData)
