+++ personalization-sdk/src/main/kotlin/com/personalization/features/notification/service/NotificationService.kt
@@ -7,12 +7,12 @@ import android.widget.Toast
 import com.personalization.di.AppModule
 import com.personalization.di.DaggerSdkComponent
 import com.personalization.features.notification.domain.model.NotificationConstants.CURRENT_IMAGE_INDEX
-import com.personalization.features.notification.domain.model.NotificationData
 import com.personalization.features.notification.presentation.helpers.NotificationHelper
 import com.personalization.features.notification.presentation.helpers.NotificationImageHelper
 import com.personalization.features.notification.service.mapper.toNotificationData
 import com.personalization.resources.NotificationResources.NOTIFICATION_LOADING_DATA_ERROR
 import com.personalization.resources.NotificationResources.NOTIFICATION_LOADING_IMAGE_ERROR
+import com.personalization.sdk.data.models.dto.notification.NotificationData
 import java.io.IOException
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
